requirements.txt
Django==4.2.1
djangorestframework==3.14.0
celery==5.2.7
python-decouple==3.4
redis==3.5.3
